pyproject.toml
[tool.poetry]
name = "bells-leaderboard-mock-up"
version = "0.1.0"
description = ""
authors = ["brash6 <hadrienmariaccia@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
streamlit = "^1.41.0"
pandas = "^2.2.3"
plotly = "^5.24.1"
openai = "^1.57.2"
anthropic = "^0.40.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
